Improve parallelism in implementation
Improve the parallelism in t he declarations in the implementation to clarify
that the difference in the overloads is due to the 'Other' parameter for
removing it from consideration for overload since the U parameter is *already*
a reference. Checking for 'const U&' is no longer the right thing to check.